Rename qt layers
