+++ napari/_qt/qt_base_layer.py
